create device table component, add flashing effects to recently changed values
