+++ src/components/TableDeviceController.tsx
@@ -1,0 +1,88 @@
+import { useMemo, useEffect, useState } from "react";
+
+import useOphydSocket from "@/hooks/useOphydSocket";
+import { Table, TableHeader, TableHead, TableCaption, TableBody, TableRow, TableCell } from "./ui/table";
+import ControllerAbsoluteMove from "./ControllerAbsoluteMove";
+import ControllerRelativeMove from "./ControllerRelativeMove";
+
+export default function TableDeviceController() {
+    const wsUrl = useMemo(()=>'ws://localhost:8000/ophydSocket', []);
+    const deviceNameList = useMemo(()=>['IOC:m1', 'IOC:m2', 'IOC:m3'], []);
+
+    const { devices, handleSetValueRequest, toggleDeviceLock, toggleExpand } = useOphydSocket(wsUrl, deviceNameList);
+
+    // State to track flashing rows
+    const [flashingRows, setFlashingRows] = useState<Record<string, boolean>>({});
+
+    useEffect(() => {
+        const updatedFlashingRows: Record<string, boolean> = {};
+        const currentTime = Date.now() / 1000; // Current time in seconds
+
+        Object.keys(devices).forEach((deviceName) => {
+            const device = devices[deviceName];
+            if (device.timestamp && currentTime - device.timestamp <= 0.03) {
+                updatedFlashingRows[deviceName] = true;
+
+                // Remove the flash effect after 1 second timeout, this timeout needs to match the animation duration of the tailwind class to avoid flutter effect
+                setTimeout(() => {
+                    setFlashingRows((prev) => ({
+                        ...prev,
+                        [deviceName]: false,
+                    }));
+                }, 500);
+            }
+        });
+
+        setFlashingRows(updatedFlashingRows);
+    }, [devices]);
+    return (
+        <Table className="max-w-[1000px] m-auto bg-neutral-50 border border-neutral-200" >
+        <TableCaption>Ophyd Devices.</TableCaption>
+        <TableHeader>
+            <TableRow>
+            <TableHead className="w-64">Device Name</TableHead>
+            <TableHead className="text-center">Current Value</TableHead>
+            <TableHead className="text-left">Absolute Move</TableHead>
+            <TableHead className="text-center">Relative Move</TableHead>
+            </TableRow>
+        </TableHeader>
+            <TableBody>
+                {
+                    Object.keys(devices).map((deviceName) => {
+                        const device = devices[deviceName];
+                        return (
+                            <TableRow key={deviceName} className={`${flashingRows[deviceName] ? 'animate-flash1' : ''}`}>
+                                <TableCell 
+                                    className="hover:cursor-pointer py-5" 
+                                    onClick={()=>toggleExpand(deviceName, device.locked)}
+                                >
+                                    <>
+                                        <p>{deviceName}</p>
+                                        {device.expanded && <pre className="text-xs">{JSON.stringify(device, null, 2)}</pre>}
+                                    </>
+                                </TableCell>
+                                <TableCell className="text-center text-md">
+                                    {`${device.value} ${device.units ? device.units.slice(0, 3) : 'n/a'}`}
+                                </TableCell>
+                                <TableCell className="">
+                                    <ControllerAbsoluteMove 
+                                        handleEnter={(input)=>input!==null && handleSetValueRequest(deviceName, input)} 
+                                        inputLabel={device.units && device.units.slice(0,3)}
+                                    />
+                                </TableCell>
+                                <TableCell className="">
+                                    <ControllerRelativeMove
+                                        className="justify-center" 
+                                        handleEnter={(input)=>input!==null && handleSetValueRequest(deviceName, input)} 
+                                        inputLabel={device.units && device.units.slice(0,3)} 
+                                        currentValue={typeof device.value === 'number' ? device.value : null}
+                                    />
+                                </TableCell>
+                            </TableRow>
+                        )
+                    })
+                }
+            </TableBody>
+        </Table>
+    )
+}
